+++ src/widgets/HeaderCopy/TutorHeader/index.tsx
@@ -1,38 +1,12 @@
-import React from 'react';
-
-import { useLocation, useNavigate } from 'react-router';
-
 import AuthHeader from '../AuthHeader';
+import useHeaderConfig from '../useHeaderConfig';
 
 import { CommonHeaderProps } from '../types';
 
-const TutorHeader: React.FC<CommonHeaderProps> = ({ onLogout }) => {
-  const navigate = useNavigate();
-  const location = useLocation();
-  return (
-    <AuthHeader
-      navItems={[
-        { text: 'Анкета', path: '/application' },
-        { text: 'Мои заявки', path: '/tutor-requests' },
-        { text: 'Личный кабинет', path: '/profile' }
-      ]}
-      avatarMenuItems={[
-        {
-          text: 'Сменить аккаунт',
-          icon: '/assets/icons/change.svg',
-          onClick: () => {
-            navigate('/login', { state: { backgroundLocation: location } });
-            onLogout();
-          }
-        },
-        {
-          text: 'Выход',
-          icon: '/assets/icons/exit.svg',
-          onClick: onLogout
-        }
-      ]}
-    />
-  );
+const TutorHeader = ({ role, onLogout }: CommonHeaderProps) => {
+  const config = useHeaderConfig({ role, onLogout });
+  return <AuthHeader {...config} />;
 };
 
 export default TutorHeader;
+
